requirements.txt
alembic==1.0.2
astroid==2.0.4
autopep8==1.4.3
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
decorator==4.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
mccabe==0.6.1
psycopg2==2.7.6.1
pycparser==2.19
pylint==2.1.1
python-dateutil==2.7.5
python-editor==1.0.3
python-slugify==1.2.6
six==1.11.0
SQLAlchemy==1.2.14
SQLAlchemy-Utils==0.33.6
typed-ast==1.1.0
Unidecode==1.0.22
validators==0.12.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
WTForms-Alchemy==0.16.7
WTForms-Components==0.10.3
